NumberGuessingGame.py
In this game user is asked to guess a number between user entered range and it will be compared to a computer generated number. If correct, score will be given. If not, user will have 2 more chances to play the game. When the user will fail all chances the game exits and the score gets printed in the console.